requirements.txt
Flask==2.2.3
Flask-Mail==0.9.1
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
python-dotenv==1.0.0
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.2
joblib==1.2.0
requests==2.28.2
gunicorn==20.1.0 